abstracted out the multitask data collator from the roberta specific code
